test: consolidate live-query optimistic scenarios
